+++ RELEASE_NOTES.md
@@ -15,7 +15,8 @@ This release adds a lot of new `RichTextAction` cases and adjusts the context an
 * `RichTextKeyboardToolbar` has a new config to always be shown.
 * `RichTextView` has a new theme that lets you define its style.
 * `RichTextViewComponent` has a new `hasRichTextStyle` function.
-* `RichTextViewComponent` has a new `toggleRichTextStyle` function. 
+* `RichTextViewComponent` has a new `toggleRichTextStyle` function.
+* `RichTextViewComponent` now handles superscript changes properly. 
 
 ### 💡 Adjustments